fix: apply dompurify to text of linkItem only on client rendering
